Move React-specific code to separate module
